Update API to properly handle array parameters / returns
Using the emscripten 'array' parameter type only works with u8 values.
If values larger than that are passed into the wrapped function, you get
garbled values on the other side. One could work around this, by
converting the input array into its u8 array representation and then do
the opposite on the rust side. However, we can make this completely
transparent for the rust side, by instead allocation the buffers
ourselves on the JS side, ignoring emscripten.

For return values, this is effectively the same, just vice-versa. Rust
already allocates the necessary memory and on the JS side we just need
to treat the data depending on the size of the elements. We do copy the
elements here so we can free the wasm memory region quickly to avoid any
memory leakage, as rust needs to let go of the memory control here. For
now, this copy is fine and does not introduce problems, but we might
have to revisit this in the future.