Add shorthand for getting values without specifying type
On the rust side, we export a function for each of the number types that
can be returned. This is necessary to properly handle the transfer to js
as they will be in different buffers, depending on their type. When it
comes to actually using the values, we don't really care as they're just
a number in JS so the distinction doesn't make sense after the transfer.
This now provides type agnostic functions to get the values from rust,
making it also easier in cases where we cannot be sure what type it is
beforehand (e.g. `f32` or `f64`).