Update to emscripten 2.0.9
Closes #3.
Starting with emscripten v2.0.2 it was now required to use `--no-entry`
when building a webassembly module that had no `main` function. This is
the case for us as we use it more as a library than a standalone
program. By applying that flag to the build, this now builds
successfully for wasm.