Build on top of head-scatcher
This builds the wasm-js bridge on top of head-scratcher with wasm-pack instead
of the emscripten build of netcdf-rs. Wasm-pack provides easier interop between
 rust and js which makes the interaction itself easier as well. We now provide
an implementation for how files and web resources work on the web to
head-sratcher. These need to be sync as the `Read` and `Seek` traits are sync
as well which results in the requirement to be run in a web worker.

The api is no longer necessary as many of the functions are now provided by
wasm-pack (e.g. allocation) and as a result has been removed.

To build, one now needs to run `wasm-pack build --target no-modules` instead of
 just `cargo build`. We need `wasm-pack` now to create the necessary wrappers
and provide us with a complete package. Notably, we need `no-modules` as target
 because it will be run inside a web worker which cannot use modules.