Update docker and build script
Given the recent changes and migration away from emscripten, the
buildscript needs to be adjusted to reflect those changes. The
instructions have been gotten much simpler as a result of the migration
as it's a rust only process now. The Dockerfile was equally adjusted as
we no longer require the emscripten toolchain and can use the rust base
instead.