Handle errors now exposed by head-scratcher
head-scratcher previously panicked on errors, so we didn't have to deal with
any potential issues. It now converts the errors to javascript which will
result in a proper exception on that side.

This did require duplicating the error codes for two reasons. First to not
pollute the head-scratcher with serialization dependencies that it otherwise
does not need. Secondly, some values contained in the error are not
serializable. The former could be worked around using features, however the
latter would create more and more complexity, instead of doing it here with a
very straight forward way of creating a matching error type just with inner
types that are serializable.